backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.5
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.15.2
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
google-auth==2.38.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msgpack==1.1.0
nodejs==0.1.1
npm==0.1.1
numpy==2.2.3
optional-django==0.1.0
pandas==2.2.3
perl==1.0.0
psycopg==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
zope.interface==7.2
